separate spire and tornjak APIs
